CORE-5114 Move integration tests into a test package to avoid package name clash somewhere buried deeply
